Added Pizza detail page, return to home function

+++ Pizzeria_website/pizzas/models.py
@@ -4,8 +4,9 @@ from django.db import models
 class Pizza(models.Model):
     name = models.CharField(max_length=50)
     price = models.PositiveIntegerField()
-    score = models.PositiveIntegerField()
+    score = models.FloatField()
     details = models.TextField(max_length=300)
+
 
     def __str__(self):
         return f"Pizza {self.name}"
